+++ app/Http/Controllers/Admin/SubCategoryController.php
@@ -3,10 +3,11 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\SubCategoryRequest;
+use App\Http\Requests\SubCategory\AddSubCategoryRequest;
+use App\Http\Requests\SubCategory\UpdateSubCategoryRequest;
 use App\Http\Resources\SubCategoryResource;
 use App\Models\Category;
-use Illuminate\Http\Request;
+use App\Models\CategoryProduct;
 use Illuminate\Support\Facades\Storage;
 
 class SubCategoryController extends Controller
@@ -19,8 +20,8 @@ class SubCategoryController extends Controller
     public function index()
     {
         // return Config::get('app.locale');
-       $categories = SubCategoryResource::collection(Category::where('is_parent','0')->get());
-        return view('admin.pages.sub-categories.index',['categories'=>$categories]);
+        $categories = SubCategoryResource::collection(Category::where('is_parent', 0)->get());
+        return view('admin.pages.sub-categories.index', ['categories' => $categories]);
     }
 
     /**
@@ -28,9 +29,10 @@ class SubCategoryController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create() {
-        $categories = Category::where('is_parent','1')->get(['id','name_en','name_ar']);
-        return view('admin.pages.sub-categories.create',['categories'=>$categories]);
+    public function create()
+    {
+        $categories = Category::where('is_parent', 1)->get(['id', 'name_en', 'name_ar']);
+        return view('admin.pages.sub-categories.create', ['categories' => $categories]);
     }
 
     /**
@@ -39,34 +41,36 @@ class SubCategoryController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(SubCategoryRequest $request) {
+    public function store(AddSubCategoryRequest $request)
+    {
+        // dd($request['name_en']);
         try {
             if ($request->has('image')) {
                 $image_path = upload_image('subcategory', $request->image);
                 Category::create([
-                    'name_en'=> $request['name_en'],
-                    'name_ar'=> $request['name_ar'],
-                    'status'=> $request->has('status')? 1:0,
-                    'is_parent'=> 0,
-                    'parent_id'=> $request['parent_id'],
+                    'name_en' => $request['name_en'],
+                    'name_ar' => $request['name_ar'],
+                    'status' => $request->has('status') ? 1 : 0,
+                    'is_parent' => 0,
+                    'parent_id' => $request['parent_id'],
                     'slug' => str_slug($request['name_en']),
-                    'image'=> $image_path,
+                    'image' => $image_path,
                 ]);
-                return redirect()->route('admin.subcategory')->with('success',"Category Added Successfully");
-            }else {
+                return redirect()->route('admin.subcategory')->with('success', "Category Added Successfully");
+            } else {
                 Category::create([
-                    'name_en'=> $request['name_en'],
-                    'name_ar'=> $request['name_ar'],
-                    'status'=> $request->has('status')? 1:0,
-                    'is_parent'=> 0,
-                    'parent_id'=> $request['parent_id'],
+                    'name_en' => $request['name_en'],
+                    'name_ar' => $request['name_ar'],
+                    'status' => $request->has('status') ? 1 : 0,
+                    'is_parent' => 0,
+                    'parent_id' => $request['parent_id'],
                     'slug' => str_slug($request['name_en']),
                 ]);
-                return redirect()->route('admin.subcategory')->with('success',"Category Added Successfully");
+                return redirect()->route('admin.subcategory')->with('success', "Category Added Successfully");
             }
         } catch (\Throwable $th) {
             return $th;
-            return redirect()->back()->with('error',"sorry.. cannot add Category right now! please try again later");
+            return redirect()->back()->with('error', "sorry.. cannot add Category right now! please try again later");
         }
     }
 
@@ -89,7 +93,7 @@ class SubCategoryController extends Controller
      */
     public function active($slug)
     {
-        return 'finishing sub cat and product first';
+        // return 'finishing sub cat and product first';
         $category = Category::whereSlug($slug)->firstOrFail();
         $category->status = 1;
         $category->update();
@@ -97,15 +101,18 @@ class SubCategoryController extends Controller
     }
     public function unActive($slug)
     {
-        return 'finishing sub cat and product first';
+        // return 'finishing sub cat and product first';
         $category = Category::whereSlug($slug)->firstOrFail();
         $category->status = 0;
         $category->update();
         return redirect()->route('admin.subcategory')->with('success', 'The "' . $category->name_en . '" Category status has been Unactivated Successfuly');
     }
-    public function edit($id)
+    public function edit($slug)
     {
-        //
+        $category = Category::whereSlug($slug)->firstOrFail();
+        $categories = Category::where('is_parent', 1)->get(['id', 'name_en', 'name_ar']);
+
+        return view('admin.pages.sub-categories.edit', get_defined_vars());
     }
 
     /**
@@ -115,9 +122,34 @@ class SubCategoryController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(UpdateSubCategoryRequest $request, $slug)
     {
-        //
+        $category = Category::whereSlug($slug)->firstOrFail();
+        // dd($category->image);
+        $image_path = $category->image;
+        if ($request->hasFile('image')) {
+            drop_image($image_path);
+            $image_path = upload_image('subcategory', $request->image);
+        }
+        $new_slug = $category->slug;
+        if ($category->name_en !== $request['name_en']) {
+            $new_slug = str_slug($request['name_en']);
+            $count = 1;
+            while (Category::whereSlug($new_slug)->exists()) {
+                $new_slug = str_slug($request['name_en']) . "-" . $count;
+                $count++;
+            }
+        }
+        $category->update([
+            'name_en' => $request['name_en'],
+            'name_ar' => $request['name_ar'],
+            'status' => $request->has('status') ? 1 : 0,
+            'is_parent' => 0,
+            'parent_id' => $request['parent_id'],
+            'slug' => $new_slug,
+            'image' => $image_path,
+        ]);
+        return redirect()->route('admin.subcategory')->with('success', 'The "' . $category->name_en . '" Category has been Updated Successfuly');
     }
 
     /**
@@ -128,14 +160,14 @@ class SubCategoryController extends Controller
      */
     public function destroy($slug)
     {
-        return 'finishing sub cat and product first';
-        $category = Category::findOrFail($slug);
-        if ($category->image != 'notfound.jpg') {
-            Storage::delete('/public/assets/images/subcategory/' . $category->image);
+        // dd()
+        // return 'finishing sub cat and product first';
+        $category = Category::where(['slug' => $slug, 'is_parent' => 0])->firstOrFail();
+        if (CategoryProduct::where('category_id',$category->id)->exists()) {
+            return redirect()->back()->with('error', 'Sorry.. Cannot delete this Category!! There are products belongs to');
         }
+        drop_image($category->image);
         $category->delete();
-
         return redirect()->route('admin.subcategory')->with('success', 'The ' . $category->name_en . ' Category has been deleted successfully');
-
     }
 }
